refactor - message details type, list of messages table setup
